refactor(cli): extract stale-issues classifier into shared module (#394 follow-up)
PRs #399 (doctor hygiene check, scope 1) and #401 (list-stale-issues CLI,
scope 3) shipped on independent branches, each carrying its own copy of
the classifier + fetch + thresholds + digest title regexes. With both now
merged, this is the dedupe pass.

New `packages/cli/src/stale-issues.ts`:
- Single `classifyStaleIssues()` returning the flat `StaleIssue[]` shape
  (was scope 3's API; doctor's old two-bucket return is now built on top)
- `partitionByReason()` helper for doctor's by-age-vs-digest summary view
- Single `fetchOpenIssues()` taking a `userAgent` parameter so doctor and
  list-stale-issues remain distinguishable in server-side logs
- `DIGEST_TITLE_PATTERNS`, `DEFAULT_AGE_DAYS`, `DEFAULT_SILENCE_DAYS`
  centralized

doctor.ts and list-stale-issues.ts now import from `./stale-issues.js`.
The classifier unit tests moved to `stale-issues.test.ts` (canonical
location); `doctor.test.ts` keeps only the integration tests that
exercise the hygiene-category code path. Net: -271 lines.

Behaviour is unchanged. Hand-verified both commands still work against
the hello-world example (graceful skip on local-mode).
